web/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pydantic>=2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
